ISP/pipeline.py
import tkinter as tk
from tkinter import filedialog, Label, Scale, Button, HORIZONTAL
from PIL import Image, ImageTk
import numpy as np
import cv2

# Function stubs for image processing steps
def demosaic_edge_based(raw_image, pattern='GRBG'):
    bayer_patterns = {
        'GRBG': cv2.COLOR_BAYER_GR2BGR,
        'RGGB': cv2.COLOR_BAYER_BG2BGR,
        'BGGR': cv2.COLOR_BAYER_RG2BGR,
        'GBRG': cv2.COLOR_BAYER_GB2BGR,
    }
    bayer_pattern = bayer_patterns[pattern]
    rgb_image = cv2.cvtColor(raw_image, bayer_pattern)
    return (rgb_image / 16).astype(np.uint8)

def white_balance(image):
    avg_rgb = np.mean(image, axis=(0, 1))
    gray_world_scale = np.mean(avg_rgb) / avg_rgb
    for i in range(3):
        image[:, :, i] = np.clip(image[:, :, i] * gray_world_scale[i], 0, 255)
    return image

def denoise(image, blur_size=5):
    if blur_size % 2 == 0:
        blur_size += 1
    return cv2.GaussianBlur(image, (blur_size, blur_size), 0)

def gamma_correction(image, gamma=2.2):
    inv_gamma = 1.0 / gamma
    lookup_table = np.array([(i / 255.0) ** inv_gamma * 255 for i in range(256)]).astype('uint8')
    return cv2.LUT(image, lookup_table)

def sharpen(image, amount=1.5, blur_size=5):
    blurred = cv2.GaussianBlur(image, (blur_size, blur_size), 0)
    return cv2.addWeighted(image, 1 + amount, blurred, -amount, 0)

class ISPApp:
    def __init__(self, root):
        self.root = root
        self.root.title("ISP App")

        # Initialize attributes
        self.raw_image = None
        self.processed_image = None

        # Load and Save Buttons
        load_button = Button(root, text="Load Image", command=self.load_image)
        load_button.pack(pady=5)

        save_button = Button(root, text="Save Image", command=self.save_image)
        save_button.pack(pady=5)

        # Sliders for adjusting ISP parameters
        self.gamma_slider = Scale(root, from_=1.0, to=3.0, resolution=0.1, orient=HORIZONTAL, label="Gamma Correction")
        self.gamma_slider.pack(fill='x', padx=10, pady=5)

        self.sharpen_slider = Scale(root, from_=0, to=5, resolution=0.1, orient=HORIZONTAL, label="Sharpening Amount")
        self.sharpen_slider.pack(fill='x', padx=10, pady=5)

        self.blur_slider = Scale(root, from_=1, to=11, resolution=2, orient=HORIZONTAL, label="Blur Kernel Size")
        self.blur_slider.pack(fill='x', padx=10, pady=5)

        self.sharpen_blur_slider = Scale(root, from_=1, to=11, resolution=2, orient=HORIZONTAL, label="Sharpen Blur Kernel Size")
        self.sharpen_blur_slider.pack(fill='x', padx=10, pady=5)

        # Process and Update Buttons
        process_button = Button(root, text="Process Image", command=self.update_image)
        process_button.pack(pady=10)

        # Label to display images
        self.image_label = Label(root)
        self.image_label.pack()

    def load_image(self):
        # Load the RAW image file
        file_path = filedialog.askopenfilename()
        if file_path:
            self.raw_image = np.fromfile(file_path, dtype=np.uint16).reshape((1280, 1920))
            self.update_image()

    def update_image(self):
        if self.raw_image is None:
            return

        # Apply ISP processing steps
        demosaiced_image = demosaic_edge_based(self.raw_image)
        white_balanced_image = white_balance(demosaiced_image)
        blur_size = self.blur_slider.get()
        denoised_image = denoise(white_balanced_image, blur_size)

        # Apply Gamma Correction and Sharpening based on slider values
        gamma_value = self.gamma_slider.get()
        gamma_corrected_image = gamma_correction(denoised_image, gamma_value)

        sharpen_amount = self.sharpen_slider.get()
        sharpen_blur_size = self.sharpen_blur_slider.get()
        self.processed_image = sharpen(gamma_corrected_image, sharpen_amount, sharpen_blur_size)

        # Convert processed image to displayable format
        display_image = Image.fromarray((self.processed_image).astype(np.uint8))  # Ensure 8-bit format
        display_image = display_image.resize((640, 480))  # Resize for GUI display

        # Update image on GUI
        self.tk_image = ImageTk.PhotoImage(display_image)
        self.image_label.config(image=self.tk_image)

    def save_image(self):
        # Save the processed image
        if self.processed_image is not None:
            save_path = filedialog.asksaveasfilename(defaultextension=".png",
                                                     filetypes=[("PNG files", "*.png"), ("All files", "*.*")])
            if save_path:
                Image.fromarray((self.processed_image).astype(np.uint8)).save(save_path)

# Main application loop
root = tk.Tk()
app = ISPApp(root)
root.mainloop()